+++ src/main/java/net/fabricmc/mappingio/format/jobf/JobfFileReader.java
@@ -32,6 +32,9 @@ import net.fabricmc.mappingio.tree.MemoryMappingTree;
 
 /**
  * {@linkplain MappingFormat#JOBF_FILE JOBF file} reader.
+ *
+ * <p>Crashes if a second visit pass is requested without
+ * {@link MappingFlag#NEEDS_MULTIPLE_PASSES} having been passed beforehand.
  */
 public class JobfFileReader {
 	private JobfFileReader() {
@@ -93,13 +96,13 @@ public class JobfFileReader {
 						// field: f <cls-a>.<name-a>:<desc-a> = <name-b>
 						// method: m <cls-a>.<name-a><desc-a> = <name-b>
 						String src = reader.nextCol();
-						if (src == null || src.isEmpty()) throw new IOException("missing class/name/desc a in line "+reader.getLineNumber());
+						if (src == null || src.isEmpty()) throw new IOException("missing class-/name-/desc-a in line "+reader.getLineNumber());
 
 						int nameSepPos = src.lastIndexOf('.');
-						if (nameSepPos <= 0 || nameSepPos == src.length() - 1) throw new IOException("invalid class/name/desc a in line "+reader.getLineNumber());
+						if (nameSepPos <= 0 || nameSepPos == src.length() - 1) throw new IOException("invalid class-/name-/desc-a in line "+reader.getLineNumber());
 
 						int descSepPos = src.lastIndexOf(isField ? ':' : '(');
-						if (descSepPos <= 0 || descSepPos == src.length() - 1) throw new IOException("invalid name/desc a in line "+reader.getLineNumber());
+						if (descSepPos <= 0 || descSepPos == src.length() - 1) throw new IOException("invalid name-/desc-a in line "+reader.getLineNumber());
 
 						readSeparator(reader);