+++ CHANGELOG.md
@@ -4,10 +4,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added IntelliJ IDEA migration map reader and writer
+- Added `MappingFormat#features()` to allow for more fine-grained programmatic querying of format capabilities
 - Added tests to validate our writer outputs against 3rd-party readers
 - Overhauled the internal `ColumnFileReader` to behave more consistently
 - Made handling of the `NEEDS_MULTIPLE_PASSES` flag more consistent, reducing memory usage in a few cases
 - Made some internal methods in Enigma and TSRG readers actually private
+- Added missing `visitElementContent` calls to CSRG and Recaf Simple readers
+- Fixed member mapping merging via tree-API in `MemoryMappingTree`
 
 ## [0.6.1] - 2024-04-15
 - Fixed CSRG and JAM writers sometimes skipping elements whose parents have incomplete destination names